nfc: Allow wallet role holder to declare associated package name
Reuse existing property to allow the wallet role holder to declare the
package name of the associated role holder. If the property is set to
true, then it defaults back to the existing behavior of matching apps
with the same signature.

Also, allow role holder to set the property either at the application
level or HCE service component level.

Bug: 441268440
Test: Manual verification using gmscore changes
Test: atest CtsNfcTestCases CtsNfcHceMultiDeviceTestCases
Flag: EXEMPT BUGFIX
Change-Id: I8535f828b27713111ac9ce7a770d7af1cefeff7b